Progress on tracking objectives and their state
Track the random objectives (both specified and not) separately from the
static ones, and the completion status in a separate set, so we can regenerate
the objectives whenever the flagstring changes without losing state.

Stop carrying a whole Location.Metadata (now Context) in the main Model: we're
interested in all its contents, so break them out to the top level of the Model
and construct the Context just when we need it to call Location methods.