Define locations with a list of Values to omit a bunch of Nothings and 0s
Keeping the same individual characters/bosses/keyItem fields in the Data
record, as it makes more sense to have those fields than to mess around with
definining them dynamically (e.g. a Set or Dict of Values, the querying of
which has to permit for more uncertainty or "possible" impossibilities).