Replace the binary 'checked' Location property with a trinary Status
This lets us have a separate "seen but not done with it yet" state,
e.g. for shops we may want to come back to, or locations where we've
peeked at the boss and opted out without actually completing it.

Currently setting the Seen status on right-click; may need to revisit
to make this more obvious or mobile-friendly.