Properly gate both Sheila checks on visiting Sylph Cave
Switched ToggleLocationStatus to carry a whole Location rather than just a
Key to avoid an impossible Maybe. Should be that nothing else can update that
Location in the model in the interim, so it's fine that we're using the value
we were passed rather than updating what's in the dictionary. Probably.

Turns out it's important not to call Dict.insert on the locations dictionary,
as it messes with the order.