Replace elm-bootstrap dropdowns with datalist-backed inputs
Using the native browser datalist input gets us type-to-filter while reducing
the amount of low-level UI we have to implement ourselves. This also lets us
remove the elm-bootstrap dependency and its styles.
On the downside, we have very limited ability to style the inputs, and the UX
may differ between browsers in unexpected/undesirable ways.